% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APLA_Liberalisation}
\alias{APLA_Liberalisation}
\title{APLA_Liberalisation}
\format{
A data frame with 552 rows and 71 variables:
\describe{
  \item{\code{Q1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q2}}{double COLUMN_DESCRIPTION}
  \item{\code{Q200#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q202#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q204#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q278#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q280#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q133#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q274#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q188#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q170#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q172#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q174#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q180#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q184#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q101#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q103#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q107#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q113#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q109#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q253#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q208#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q145#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q190#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q194#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q320#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q54#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Q58#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA1#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA3#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA5#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA7#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA9#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA11#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA13#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA15#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA17#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA19#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA23#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA27#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA29#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA31#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA33#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA35#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA37#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA39#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA41#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA43#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA45#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA47#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA49#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA51#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA53#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA55#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA57#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA59#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA61#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA63#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{LA65#2_1}}{double COLUMN_DESCRIPTION}
  \item{\code{...60}}{logical COLUMN_DESCRIPTION}
  \item{\code{TOTAL1}}{double COLUMN_DESCRIPTION}
  \item{\code{TOTAL0}}{double COLUMN_DESCRIPTION}
  \item{\code{...63}}{logical COLUMN_DESCRIPTION}
  \item{\code{Total_10}}{double COLUMN_DESCRIPTION}
  \item{\code{...65}}{logical COLUMN_DESCRIPTION}
  \item{\code{Ratio1toTotal}}{double COLUMN_DESCRIPTION}
  \item{\code{...67}}{logical COLUMN_DESCRIPTION}
  \item{\code{Liberalisation}}{double COLUMN_DESCRIPTION}
  \item{\code{Formula0stoTotalinYear}}{logical COLUMN_DESCRIPTION}
  \item{\code{Restrictiveness}}{double COLUMN_DESCRIPTION}
  \item{\code{Formula:1stoTotalinaYear}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
APLA_Liberalisation
}
\description{
APLA Liberalisation  include the files where the calculation
of the two dependent variables was conducted
}
\keyword{datasets}
