<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APLA_Map}
\alias{APLA_Map}
\title{APLA_Map}
\format{
A data frame with 667 rows and 34 variables:
\describe{
  \item{\code{Country}}{Country Codified}
  \item{\code{Year}}{Year Codified}
  \item{\code{Codified}}{Integer [0,1] 1 for Inclusion of country into the APLA Dataset, 0 otherwise}
}
}
\source{
\url{https://github.com/HammoudG/APLA_Dataset}
}
\usage{
APLA_Map
}
\description{
APLA_Map is slightly modified versions of the dataset Data_APLA,
 which allows to better  map certain variables.
}
\keyword{datasets}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APLA_Map}
\alias{APLA_Map}
\title{APLA_Map}
\format{
A data frame with 667 rows and 34 variables:
\describe{
  \item{\code{Country}}{Country Codified}
  \item{\code{Year}}{Year Codified}
  \item{\code{Codified}}{Integer [0,1] 1 for Inclusion of country into the APLA Dataset, 0 otherwise}
}
}
\source{
\url{https://github.com/HammoudG/APLA_Dataset}
}
\usage{
APLA_Map
}
\description{
APLA_Map is slightly modified versions of the dataset Data_APLA,
 which allows to better  map certain variables.
}
\keyword{datasets}
>>>>>>> 0939008be34f3170a5d71fbc1fbe7379900bd9a3
