% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APLA_Regulatory_Complexity}
\alias{APLA_Regulatory_Complexity}
\title{APLA_Regulatory_Complexity}
\format{
A data frame with 552 rows and 135 variables:
\describe{
  \item{\code{Q1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q2}}{character COLUMN_DESCRIPTION}
  \item{\code{Q5#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q5#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q7#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q7#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q9#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q9#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q11#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q11#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q13#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q13#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q15#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q15#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q17#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q17#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q19#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q19#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q200#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q200#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q202#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q202#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q204#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q204#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q278#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q278#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q280#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q280#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q133#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q133#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q274#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q274#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q188#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q188#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q170#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q170#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q172#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q172#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q174#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q174#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q180#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q180#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q184#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q184#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q101#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q101#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q103#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q103#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q107#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q107#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q113#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q113#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q109#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q109#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q253#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q253#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q208#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q208#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q145#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q145#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q190#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q190#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q194#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q194#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q320#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q320#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q54#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q54#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q58#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Q58#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA1#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA1#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA3#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA3#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA5#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA5#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA7#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA7#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA9#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA9#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA11#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA11#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA13#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA13#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA15#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA15#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA17#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA17#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA19#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA19#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA23#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA23#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA27#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA27#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA29#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA29#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA31#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA31#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA33#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA33#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA35#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA35#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA37#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA37#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA39#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA39#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA41#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA41#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA43#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA43#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA45#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA45#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA47#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA47#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA49#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA49#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA51#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA51#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA53#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA53#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA55#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA55#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA57#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA57#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA59#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA59#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA61#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA61#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA63#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA63#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA65#2_1}}{character COLUMN_DESCRIPTION}
  \item{\code{LA65#3_1_1}}{character COLUMN_DESCRIPTION}
  \item{\code{Art}}{double COLUMN_DESCRIPTION}
  \item{\code{Total}}{double COLUMN_DESCRIPTION}
  \item{\code{Percentage_Art}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
APLA_Regulatory_Complexity
}
\description{
and APLA Regulatory Complexity includes the files where the calculation of
the two dependent variables was conducted
}
\keyword{datasets}
