% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_APLA_1}
\alias{Data_APLA_1}
\title{Data_APLA_1}
\format{
A data frame with 580 rows and 33 variables:
\describe{
  \item{\code{Country}}{character COLUMN_DESCRIPTION}
  \item{\code{Year}}{double COLUMN_DESCRIPTION}
  \item{\code{Art}}{double COLUMN_DESCRIPTION}
  \item{\code{Total}}{double COLUMN_DESCRIPTION}
  \item{\code{Regulatory_Complexity}}{double COLUMN_DESCRIPTION}
  \item{\code{Polity2}}{double COLUMN_DESCRIPTION}
  \item{\code{IntMigStock}}{double COLUMN_DESCRIPTION}
  \item{\code{Left1_Other0}}{double COLUMN_DESCRIPTION}
  \item{\code{Liberalisation}}{double COLUMN_DESCRIPTION}
  \item{\code{MigSpain}}{double COLUMN_DESCRIPTION}
  \item{\code{MigUS}}{double COLUMN_DESCRIPTION}
  \item{\code{MigSpainUS}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalPopinY}}{double COLUMN_DESCRIPTION}
  \item{\code{MigSpainUSPerc}}{double COLUMN_DESCRIPTION}
  \item{\code{GrowthGDPperCap}}{double COLUMN_DESCRIPTION}
  \item{\code{Trade_Perc_GDP}}{double COLUMN_DESCRIPTION}
  \item{\code{RefugeeAndLikeSit}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalPop}}{double COLUMN_DESCRIPTION}
  \item{\code{RefAsPerc}}{double COLUMN_DESCRIPTION}
  \item{\code{GDPperCapPPP}}{double COLUMN_DESCRIPTION}
  \item{\code{Lib100}}{double COLUMN_DESCRIPTION}
  \item{\code{x}}{double COLUMN_DESCRIPTION}
  \item{\code{y}}{double COLUMN_DESCRIPTION}
  \item{\code{CountryID}}{character COLUMN_DESCRIPTION}
  \item{\code{South_America}}{double COLUMN_DESCRIPTION}
  \item{\code{MigSpainUSLog}}{double COLUMN_DESCRIPTION}
  \item{\code{lag1}}{double COLUMN_DESCRIPTION}
  \item{\code{lag2}}{double COLUMN_DESCRIPTION}
  \item{\code{_est_est1}}{double COLUMN_DESCRIPTION}
  \item{\code{_est_est2}}{double COLUMN_DESCRIPTION}
  \item{\code{_est_est3}}{double COLUMN_DESCRIPTION}
  \item{\code{_est_est4}}{double COLUMN_DESCRIPTION}
  \item{\code{VDEM_Polyarchy}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Data_APLA_1
}
\description{
Data_APLA_1 is slightly modified versions of the dataset Data_APLA,
 which allows to better plot or chart certain variables.
}
\keyword{datasets}
