% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Refugees_UNHCR}
\alias{Data_Refugees_UNHCR}
\title{DATASET_TITLE}
\format{
A data frame with 580 rows and 5 variables:
\describe{
  \item{\code{Country}}{character COLUMN_DESCRIPTION}
  \item{\code{Year}}{integer COLUMN_DESCRIPTION}
  \item{\code{RefugeeAndLikeSit}}{integer COLUMN_DESCRIPTION}
  \item{\code{TotalPop}}{integer COLUMN_DESCRIPTION}
  \item{\code{RefAsPerc}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Data_Refugees_UNHCR
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
